[refactor] service 패키지 변경
